+++ app/workers/expire_build_instance_artifacts_worker.rb
@@ -2,10 +2,14 @@ class ExpireBuildInstanceArtifactsWorker
   include Sidekiq::Worker
 
   def perform(build_id)
-    build = Ci::Build.with_expired_artifacts.reorder(nil).find_by(id: build_id)
-    return unless build
+    build = Ci::Build
+      .with_expired_artifacts
+      .reorder(nil)
+      .find_by(id: build_id)
 
-    Rails.logger.info "Removing artifacts build #{build.id}..."
+    return unless build.try(:project)
+
+    Rails.logger.info "Removing artifacts for build #{build.id}..."
     build.erase_artifacts!
   end
 end